backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0.0
requests>=2.31.0
gspread>=6.0.0
google-api-python-client>=2.137.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
